ADD: [CHORE Fixes #166253110] create user controller
